Apache-Airflow/Movie_Quality_checks_Airflow-Snowflake-Soda/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-snowflake
astro-sdk-python==1.8.1
